Move ScalaFmtLogger to util package.
