+++ cli/src/main/scala/org/scalafmt/cli/Cli.scala
@@ -6,7 +6,8 @@ import org.scalafmt.ScalaFmt
 import org.scalafmt.ScalaStyle
 import org.scalafmt.Versions
 import org.scalafmt.internal.Debug
-import org.scalafmt.internal.ScalaFmtLogger._
+import org.scalafmt.util.ScalaFmtLogger
+import ScalaFmtLogger._
 import org.scalafmt.util.FilesUtil
 
 object Cli {
