+++ core/src/main/scala/org/scalafmt/internal/Debug.scala
@@ -1,6 +1,7 @@
 package org.scalafmt.internal
 
-import org.scalafmt.internal.ScalaFmtLogger._
+import org.scalafmt.util.ScalaFmtLogger
+import ScalaFmtLogger._
 
 import java.util.concurrent.TimeUnit
