+++ core/src/main/scala/org/scalafmt/internal/Router.scala
@@ -1,7 +1,8 @@
 package org.scalafmt.internal
 
 import org.scalafmt.Error.UnexpectedTree
-import org.scalafmt.internal.ScalaFmtLogger._
+import org.scalafmt.util.ScalaFmtLogger
+import ScalaFmtLogger._
 import org.scalafmt.internal.Policy.NoPolicy
 
 import scala.collection.mutable
