+++ core/src/main/scala/org/scalafmt/util/ScalaFmtLogger.scala
@@ -1,6 +1,8 @@
-package org.scalafmt.internal
+package org.scalafmt.util
 
-import org.scalafmt.util.PrintlnLogger
+import org.scalafmt.internal.FormatToken
+import org.scalafmt.internal.Split
+
 import scala.meta.Tree
 import scala.meta.prettyprinters.Structure
 import scala.meta.tokens.Token
