+++ core/src/test/scala/org/scalafmt/FormatExperiment.scala
@@ -1,6 +1,7 @@
 package org.scalafmt
 
-import org.scalafmt.internal.ScalaFmtLogger._
+import org.scalafmt.util.ScalaFmtLogger
+import ScalaFmtLogger._
 import org.scalafmt.util.ExperimentResult
 import org.scalafmt.util.ExperimentResult.Skipped
 import org.scalafmt.util.ExperimentResult.Success
